changed to double, copied racecar code to stick

+++ main/arduino_main.cpp
@@ -60,8 +60,8 @@ void handleThrottle(void* params) {
         }
         if (myGamepad && controlScheme && myGamepad->isConnected()) {
             auto axis = controlScheme->handleThrottle();
-            int leftSpeed = axis.first;
-            int rightSpeed = axis.second;
+            double leftSpeed = axis.first;
+            double rightSpeed = axis.second;
 
             double leftPWM = LEFT_ESC.getScaler()->scaleAxis(leftSpeed);
             double rightPWM = RIGHT_ESC.getScaler()->scaleAxis(rightSpeed);
